+++ src/calculator.h
@@ -3,11 +3,11 @@
 
 #define EXPORT __declspec(dllexport)
 
-EXPORT double add(double num_a, double num_b);
-EXPORT double sub(double num_a, double num_b);
-EXPORT double mul(double num_a, double num_b);
-EXPORT double div(double num_a, double num_b);
-EXPORT double mod(double num_a, double num_b);
-EXPORT double calculator(char operator, double num_a, double num_b);
+EXPORT float add(float num_a, float num_b);
+EXPORT float sub(float num_a, float num_b);
+EXPORT float mul(float num_a, float num_b);
+EXPORT float div(float num_a, float num_b);
+EXPORT float mod(float num_a, float num_b);
+EXPORT float calculator(char operator, float num_a, float num_b);
 
 #endif
